refac: rename usecase
