+++ lib/app/home/tabs/repositories/presenter/dependencies/repository_dependencies.dart
@@ -1,4 +1,4 @@
-import 'package:app_masterclass/app/home/tabs/repositories/domain/usecases/get_posts.dart';
+import 'package:app_masterclass/app/home/tabs/repositories/domain/usecases/get_repos.dart';
 import 'package:app_masterclass/app/home/tabs/repositories/presenter/dependencies/repo_injections.dart';
 import 'package:flutter/material.dart';
 import 'package:uno/uno.dart';
